GenAI/requirements.txt
flask>=2.3.0
google-generativeai>=0.3.0
PyMuPDF>=1.23.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
werkzeug>=2.3.0
